feat: Add DTOs for Role requests and responses
Added mapper with toRoleResponse method to serializer role data to
RoleResponse data.
